First real push
- Added waypoints for enemies
- Added enemies
- Added wave spawner
- Added tower pedestal
-Organized files

I accidentally made a second scene called MainScene (as opposed to SampleScene). I've been working in MainScene because I can't delete either of them now, but I think the only difference is the layout of objects placed in the editor. They have access to all the same scripts and such.